get ingredient list from user and return matched recipes as output

get ingredient list from user and return matched recipes as output

Usage

get_ingredients(...)

Arguments

...

the function directly asks for user input, the argument is not necessary, but it can be a vector of character that contains ingredient names

Value

a 10-row dataframe of recipes based on user input

Examples

get_ingredients("egg", "sugar", "milk")
#> egg was found in the list of ingredients.
#> sugar was found in the list of ingredients.
#> milk was found in the list of ingredients.
#> You entered: egg, sugar, milk
#> Please wait while the recipes are loading...
#>                                                     title
#> 1                             Perfect Buttermilk Pancakes
#> 2                                                 Waffles
#> 3                 Pasta Primavera With Asparagus and Peas
#> 4                                          Olive Oil Cake
#> 5                                  Overnight French Toast
#> 6                                       Sugar Cookie Bars
#> 7                                          Cinnamon Rolls
#> 8  Earl Grey Tea Cake With Dark Chocolate and Orange Zest
#> 9                                Chez Ma Tante's Pancakes
#> 10                                  Rainbow Sprinkle Cake
#>                                                                                                                                    tag
#> 1                                                                              Pancakes, Buttermilk, Cast Iron, Breakfast, Easy, Quick
#> 2                       Editors' Pick, All Purpose Flour, Egg, Freezer Friendly, Milk, Unsalted Butter, Waffle, Breakfast, Main Course
#> 3   Pastas, Asparagus, Creme Fraiche, English Pea, Fettuccine, Greek Yogurt, Sugar Snap Pea, Dinner, Lunch, Quick, Main Course, Spring
#> 4                                                 Italian, Cakes, Extra Virgin Olive Oil, Lemon Juice, Lemon Zest, Make Ahead, Dessert
#> 5                                                           Brioche, Challah, French Toast, Plan Ahead, Breakfast, Brunch, Main Course
#> 6                                                    American, Bar Cookie, Cream Cheese, Sugar Cookie, Vanilla Extract, Snack, Dessert
#> 7                        American, Pastries, Brown Sugar, Cinnamon, Yeast, Breakfast, Brunch, Dessert, Christmas, Easter, Mother's Day
#> 8  American, Cakes, Cream Cheese, Dark Chocolate, Earl Grey Tea, Mascarpone, Orange Zest, Snack, Dessert, Mother's Day, New Year's Day
#> 9                                                                    American, Pancakes, Cast Iron, Egg Yolk, Breakfast, Brunch, Quick
#> 10                                            American, Cakes, Confectioners' Sugar, Cream Cheese, Rainbow Sprinkle, Dessert, Birthday
#>                                    serving
#> 1                               10 minutes
#> 2                               35 minutes
#> 3                               20 minutes
#> 4                     1 hour, plus cooling
#> 5                    1 hour, plus chilling
#> 6                 40 minutes, plus cooling
#> 7  About 1 hour, plus resting and proofing
#> 8                    1 hour, plus chilling
#> 9                               20 minutes
#> 10               1 1/2 hours, plus cooling
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   ingredients
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    2 cups all-purpose flour 3 tablespoons sugar 1 1/2 teaspoons baking powder 1 1/2 teaspoons baking soda 1 1/4 teaspoons kosher salt 2 1/2 cups buttermilk 2 large eggs 3 tablespoons unsalted butter, melted vegetable, canola or coconut oil for the pan
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                              6 tablespoons unsalted butter, more for waffle iron 2 cups/240 grams all-purpose flour 1 tablespoon/15 grams sugar 1 teaspoon/8 grams baking powder 1 teaspoon/5 grams fine sea salt 1/2 teaspoon/3 grams baking soda 1 cup plain yogurt (or see note) 1 cup milk 4 large eggs
#> 3                                                                                                                                                 1/4 pound sugar snap peas, stems trimmed 1/2 pound asparagus, ends snapped 2 tablespoons unsalted butter 3/4 cup fresh english peas 1/4 cup thinly sliced spring onion, white part only (or use shallot) 2 garlic cloves, finely chopped 1/2 teaspoon fine sea salt, more as needed black pepper, more as needed 12 ounces fettuccine or tagliatelle, preferably fresh (see recipe) 2/3 cup grated parmigiano-reggiano, at room temperature 1/2 cup creme fraiche or whole milk greek yogurt, at room temperature 3 tablespoons finely chopped parsley 1 tablespoon finely chopped tarragon
#> 4                                                                                                                                                                                                                                                                 1 cup/240 milliliters good-quality extra-virgin olive oil, plus more for the pan 2 cups/255 grams all-purpose flour, plus more for the pan 1 teaspoon kosher salt 1 teaspoon baking powder 1/4 teaspoon baking soda 1 1/2 cups/300 grams granulated sugar, plus about 2 tablespoons for sprinkling 3 large eggs, at room temperature 1 tablespoon freshly grated lemon zest plus 2 tablespoons fresh lemon juice 1 1/4 cups/295 milliliters whole milk, at room temperature
#> 5                                                                                                                                                                                                                                    2 tablespoons unsalted butter, softened, plus more for baking dish 1 pound enriched bread, such as brioche, challah or panettone sliced 3/4-inch thick 1/4 cup golden raisins (optional) 6 large eggs 2 1/2 cups whole milk 1/2 cup heavy cream 1/3 cup light brown sugar 1/2 teaspoon ground cinnamon pinch of freshly ground nutmeg 1/2 teaspoon pure vanilla extract 1/4 teaspoon kosher salt 1/3 cup pecans, chopped (optional) confectioners' sugar for dusting (optional) maple syrup, for serving
#> 6  for the bars 1 cup/225 grams unsalted butter (2 sticks), at room temperature, plus more for greasing the pan 2 3/4 cups/350 grams all-purpose flour 1/2 teaspoon fine sea salt 1 (8-ounce/225-gram) package cream cheese, at room temperature 1 1/2 cups/300 grams granulated sugar 1 large egg 2 teaspoons vanilla extract for the frosting 6 tablespoons/85 grams unsalted butter ( 3/4 stick), at room temperature 2 cups/245 grams confectioners' sugar 1 tablespoon milk or heavy cream, plus more as needed 1 teaspoon fresh lemon juice, plus more as needed 1 teaspoon vanilla extract, plus more as needed 1/4 teaspoon fine sea salt a drop or two of gel food coloring (optional) assorted sprinkles, for decorating (optional)
#> 7                                                                                                            for the dough 1 cup whole milk 4 tablespoons sugar 2 1/4 teaspoons/7 grams active dry yeast ( 1/4-ounce envelope) 4 cups/510 grams all-purpose flour, plus more for dusting 1 1/2 teaspoons baking powder 1 1/4 teaspoons kosher salt 1 large egg, lightly beaten 6 tablespoons/85 grams unsalted butter, melted nonstick spray or vegetable oil, for the bowl for the filling and assembly 1 1/4 cups/250 grams light brown sugar 2 tablespoons ground cinnamon 1/4 teaspoon kosher salt 1/2 cup/114 grams unsalted butter (1 stick), melted 3 cups/362 grams confectioners' sugar 5 to 6 tablespoons whole milk or heavy cream
#> 8                                                                          for the frosting 3/4 cup/180 milliliters heavy cream 2 teaspoons loose earl grey tea 1/4 cup/30 grams confectioners' sugar 1/2 cup/115 grams mascarpone or softened cream cheese (see tip) for the cake 1/2 cup/115 grams unsalted butter (1 stick), at room temperature, plus more for greasing the pan 1 1/2 cups/190 grams all-purpose flour 1 tablespoon loose earl grey tea 1 teaspoon baking powder 1/2 teaspoon kosher salt 1 cup/200 grams granulated sugar 2 teaspoons freshly grated orange zest (from 1 large orange) 2 large eggs, at room temperature 1/2 cup/120 milliliters whole milk, at room temperature 1/4 cup/45 grams chopped dark chocolate
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                               1 large egg 1 egg yolk 2 1/2 tablespoons baking powder 2 tablespoons granulated sugar 1 teaspoon kosher salt 1 1/4 cups whole milk 1 cup all-purpose flour 1 cup plus 2 tablespoons clarified butter (or store-bought), melted salted butter, for serving (optional) maple syrup, for serving
#> 10              for the cake 1/2 cup/110 grams unsalted butter (1 stick), slightly softened, more for pans 1 1/2 cups/190 grams all-purpose flour, more for pans 1 teaspoon baking powder 1/4 teaspoon baking soda 1/4 teaspoon fine salt 2/3 cup/158 milliliters whole milk 1 1/2 teaspoons pure vanilla extract 1 cup/200 grams sugar 1 whole egg plus 2 egg whites 1/3 cup rainbow sprinkles (not pastel, or naturally colored) for the frosting 8 ounces/225 grams cream cheese, slightly softened 1/2 cup/110 grams unsalted butter (1 stick), slightly softened 1/8 teaspoon fine salt, more to taste 3 cups/300 grams confectioners' sugar, sifted, more to taste 1/2 teaspoon pure vanilla extract 1 cup rainbow sprinkles (see note)
#>               rating                                    time
#> 1  8792 user ratings                              10 minutes
#> 2  6279 user ratings                              35 minutes
#> 3  5355 user ratings                              20 minutes
#> 4  5152 user ratings                    1 hour, plus cooling
#> 5  4948 user ratings                   1 hour, plus chilling
#> 6  4713 user ratings                40 minutes, plus cooling
#> 7  4264 user ratings About 1 hour, plus resting and proofing
#> 8  4172 user ratings                   1 hour, plus chilling
#> 9  4030 user ratings                              20 minutes
#> 10 3911 user ratings               1 1/2 hours, plus cooling
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 instructions
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Step 1 Heat the oven to 325 degrees. Whisk flour, sugar, baking powder, baking soda and kosher salt together in a bowl. Using the whisk, make a well in the center. Pour the buttermilk into the well and crack eggs into buttermilk. Pour the melted butter into the mixture. Starting in the center, whisk everything together, moving towards the outside of the bowl, until all ingredients are incorporated. Do not overbeat (lumps are fine). The batter can be refrigerated for up to one hour. Step 2 Heat a large nonstick griddle or skillet, preferably cast-iron, over low heat for about 5 minutes. Add 1 tablespoon oil to the skillet. Turn heat up to medium-low and using a measuring cup, ladle 1/3 cup batter into the skillet. If you are using a large skillet or a griddle, repeat once or twice, taking care not to crowd the cooking surface. Step 3 Flip pancakes after bubbles rise to surface and bottoms brown, about 2 to 4 minutes. Cook until the other sides are lightly browned. Remove pancakes to a wire rack set inside a rimmed baking sheet, and keep in heated oven until all the batter is cooked and you are ready to serve. Leave a Private Note on this recipe and see it here.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Step 1 Melt butter either on the stove or in the microwave. Set aside. Step 2 In a large bowl, whisk together flour, sugar, baking powder, salt, and baking soda. In a separate bowl, whisk together yogurt, milk, melted butter, and eggs. Fold wet ingredients into dry ingredients. Step 3 Preheat a waffle iron and, using a pastry brush or paper towel, lightly coat with butter. Cook waffles (using about 1/2 cup batter per waffle) until golden and crisp. Butter the iron in between batches as needed. Serve waffles immediately as they are ready, or keep them warm in a 200 degree oven until ready to serve. If you don't have yogurt (or sour cream or buttermilk will all work), substitute another 1 cup of milk mixed with 1 tablespoon lemon juice or vinegar. Leave a Private Note on this recipe and see it here.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Step 1 Bring a large pot of heavily salted water to a boil over medium-high heat. Step 2 While the water is coming to a boil, slice snap peas and asparagus stems into 1/4-inch-thick pieces; leave asparagus tips whole. Step 3 Melt butter in a large skillet over medium-high heat. Add snap peas, asparagus, English peas and onion. Cook until vegetables are barely tender (but not too soft or mushy), 3 to 4 minutes. Stir in garlic and cook 1 minute more. Season with salt and pepper; set aside. Step 4 Drop pasta into boiling water and cook until al dente (1 to 3 minutes for fresh pasta, more for dried pasta). Drain well and transfer pasta to a large bowl. Immediately toss pasta with vegetables, Parmigiano-Reggiano, creme fraiche and herbs. Season generously with salt and pepper, if needed. Leave a Private Note on this recipe and see it here.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Step 1 Heat the oven to 375 degrees. Grease a 9-inch round cake pan using extra-virgin olive oil and line the bottom with parchment paper. Oil the parchment and flour the pan, shaking out any excess flour. Step 2 In a large bowl, whisk together the flour, salt, baking powder and baking soda. Step 3 In the bowl of an electric mixer set on high, beat the sugar, eggs and lemon zest until very thick and fluffy, about 5 minutes. With the mixer still running, slowly drizzle in the oil and beat until incorporated, another 2 minutes. Reduce speed to low, and add milk and lemon juice. Gradually add the flour mixture and beat until just combined. Transfer the batter to the prepared pan, smooth the top using an offset spatula and sprinkle the top with about 2 tablespoons sugar. Step 4 Bake the cake until a skewer inserted into the center comes out clean, 40 to 45 minutes. Transfer to a rack to cool for 20 minutes, then run a knife around the edge to release the sides of the cake from the pan. Invert the cake onto a plate and then flip it back over onto the rack to cool completely. Store leftovers in an airtight container at room temperature for up to one week. Leave a Private Note on this recipe and see it here.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Step 1 Butter a 2-quart baking dish. Butter the front sides of the slices of bread and shingle the slices upright in the dish, overlapping as necessary. Nestle the raisins evenly between the slices of bread but not on top. Step 2 In a large bowl, whisk together eggs, milk, cream, sugar, cinnamon, nutmeg, vanilla and salt until well combined. Pour the milk mixture over the bread and press the bread gently to absorb. Cover with plastic wrap and refrigerate overnight. Step 3 Heat oven to 350 degrees. Remove the plastic wrap. Sprinkle the pecans evenly over the top. Bake until puffed, golden and set in the center, 50 to 60 minutes. While baking, if the top begins to get too dark, tent with foil. Let stand 10 minutes. Dust with confectioners' sugar and serve with maple syrup. Leave a Private Note on this recipe and see it here.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Step 1 Heat oven to 350 degrees. Lightly butter a 9-by-13-inch baking pan. Line the pan with parchment paper, running it up the two long sides of the pan and letting it extend past the rim by about 2 inches. Step 2 Make the bars: In a medium bowl, whisk together the flour and salt. In a large bowl with an electric mixer, or in the bowl of a stand mixer fitted with the paddle attachment, beat butter and cream cheese on medium speed until well blended, about 1 minute. Add the granulated sugar and beat until smooth, about 1 minute. Add the egg and vanilla and beat on low speed until well combined, about 1 minute. Turn off the mixer and scrape down the bowl with a rubber spatula. Gradually add the flour mixture, mixing on low speed just until blended, about 1 minute. Step 3 Using a spatula, scrape the dough into the prepared baking pan and spread it into an even layer. Bake just until the edges are starting to turn light golden brown, and a toothpick inserted in the middle has moist crumbs, 20 to 25 minutes. (Do not overbake! The bars should be quite moist, and almost slightly underbaked in the middle.) Remove the pan from the oven and set it on a wire rack. Let cool completely. When fully cooled, remove the bars from the pan using the overhanging parchment paper. Step 4 Make the frosting: Place the butter in the bowl of a stand mixer fitted with the paddle attachment. Beat the butter on medium speed until creamy, about 1 minute. Add 1 cup confectioners' sugar, beating on low speed until fully combined, then repeat with remaining 1 cup confectioners' sugar. Add the 1 tablespoon milk or heavy cream, 1 teaspoon lemon juice, the vanilla, salt and food coloring, if using, and beat on medium speed until the frosting is light and fluffy, scraping down the sides halfway through, about 4 minutes. Add more milk if needed to thin out the mixture. Add lemon juice and vanilla to taste. Step 5 Using a spatula, spread the frosting on top, then decorate with sprinkles, if using. Cut into 20 bars and serve. Leave a Private Note on this recipe and see it here.
#> 7  Step 1 Make the dough: Warm milk in a small pot over medium heat until it's warm to the touch but not yet simmering (110 degrees). Add 2 tablespoons sugar and the yeast, whisking to dissolve and break up any clumps. Let sit until it's slightly foamy and starting to bubble, about 2 to 5 minutes. (If this doesn't happen after about 5 minutes, check the expiration date on the yeast; it might be dead.) Step 2 Meanwhile, combine flour, the remaining 2 tablespoons sugar, the baking powder and the salt in a large bowl or in the bowl of a stand mixer fitted with the paddle attachment. Mix briefly to combine. Step 3 Add milk mixture, followed by the egg and, with a wooden spoon or the mixer on low, stir to combine, just to eliminate any obviously dry or wet spots. If using the mixer, you may need to scrape the dough off the paddle. Add melted butter and continue to stir until a shaggy lump of dough forms, again scraping down the paddle and helping things along with your hands if necessary to combine. Step 4 Turn dough out onto a lightly floured work surface, or keep in the mixing bowl and change to the dough hook attachment. Knead dough until it's smooth, shiny and elastic, about 4 or 5 minutes. (You shouldn't have to add any more flour at this point, but if the dough seems especially sticky, give it a light dusting.) Step 5 Lightly grease a large bowl with nonstick spray or vegetable oil and place dough in the bowl. Cover tightly with plastic wrap and place in a warm, draft-free area. (If it's cold outside, on top of the stove is generally a good bet, as long as the burners and oven aren't on.) Let sit until the dough has about doubled in size, 1 1/2 to 2 hours. Step 6 Punch dough down to release any air that has built up. Cover tightly with plastic wrap, place in the refrigerator and chill the dough for at least an hour (and as long as overnight). Step 7 Fill the dough and shape the rolls: Once dough is chilled (it should feel firm and no longer flimsy), turn it out onto a lightly floured work surface and roll to a 16 x 10-inch rectangle. The rolled dough will be about 1/4- to 1/3-inch thick. Step 8 Combine brown sugar, cinnamon and salt in a small bowl. Brush surface of the dough with half of the melted butter and sprinkle brown sugar mixture in an even layer, patting to make sure it's sticking. Drizzle with remaining butter. Step 9 Starting at the end closest to you (the long side of the dough), roll dough up into a tight coil, pressing lightly as you roll to make sure there are no gaps between the dough and the filling. Step 10 Cut log into nine even pieces, about 1 3/4-inch thick. Lightly spray a 9 x 13-inch baking dish with nonstick spray and arrange dough in three rows of three pieces each, with the spiral facing upward and with space in between each piece. (At this point, the rolls can be covered tightly with plastic wrap and refrigerated overnight, if you'd like.) Step 11 Cover rolls lightly with plastic wrap and place in a warm, draft-free area until the rolls have puffed (they won't quite double, but they will be noticeably fluffier and closer to touching) and spring back slightly when pressed with your finger, about 1 hour. (If you refrigerated the rolls overnight, this may take 1 1/2 to 2 hours.) Step 12 Heat oven to 375 degrees. Place rolls in oven and bake until they are golden brown and fluffy and the sugar has started to bubble up around the edges, caramelizing on the sides of the buns, about 32 to 35 minutes. Step 13 Meanwhile, make the glaze: combine confectioners' sugar and milk and whisk until it is thick but can still be drizzled, like a thinner frosting. (If the glaze is still too thick, thin with more milk by the teaspoonful to get desired consistency.) Keep in mind the glaze is to be applied while buns are warm, and it will thin out upon contact, so it's best to err on the thicker side. Step 14 Once rolls are out of the oven, drizzle with the powdered sugar glaze and let cool slightly in the pan before digging in. Leave a Private Note on this recipe and see it here.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Step 1 Prepare the frosting: In a small saucepan, bring 1/2 cup/120 milliliters heavy cream to a simmer over medium-high heat. Stir in the tea, remove from the heat, cover and let stand for at least 30 minutes and up to 1 hour. Strain the tea through a fine-mesh sieve, discarding the solids, and chill the remaining cream until completely cold, at least 1 hour. Step 2 Prepare the cake: Heat the oven to 350 degrees. Butter an 8-inch round cake pan and line with parchment paper. In a medium bowl, whisk together the flour, tea, baking powder and salt. Step 3 In large bowl, beat the butter and sugar with an electric mixer on medium until light and fluffy, about 3 minutes. Add the orange zest and beat to combine. Add the eggs, one at a time, and beat until combined, scraping the sides of the bowl as necessary. Beat in the flour mixture on low, until just combined, then beat in the milk. (Don't overmix.) Add the chocolate and fold it in using a spatula. Transfer the batter to the prepared pan and smooth the top. Bake just until a toothpick comes out with moist crumbs attached, 30 to 35 minutes. Transfer to a rack to cool for about 15 minutes. Then tip the cake out onto the rack to cool completely. Step 4 To finish the frosting, add the remaining 1/4 cup/60 milliliters cream and the confectioners' sugar to the tea cream. With an electric mixer on medium, beat the cream mixture until medium-stiff peaks form, 2 to 3 minutes. Add the mascarpone and beat just until stiff peaks form. (Do not overmix.) Top the cake with the frosting to serve. Store leftovers covered in the fridge for up to 3 days; let come to room temperature before serving. If using cream cheese instead of mascarpone, because it has a thicker consistency, whip the cream cheese with the confectioners' sugar first, then add the whipped tea cream. Leave a Private Note on this recipe and see it here.
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Step 1 Whisk egg and yolk together in a medium bowl. Add baking powder, sugar and salt; whisk until smooth and fluffy. Pour in half the milk, then half the flour. Using a wooden spoon, stir to combine. Add the remaining milk and flour plus 2 tablespoons clarified butter and stir briefly just until batter comes together but is still somewhat lumpy. Step 2 Heat a large 12-inch cast-iron skillet or griddle over medium-high for at least 5 minutes. Pour about 1/4 cup clarified butter into the pan. When the surface of the clarified butter starts to shimmer, ladle about 1/3 cup of the batter into the skillet for each pancake, leaving a couple of inches between each pancake. Add more clarified butter as pancakes cook to keep about 1/8 inch of fat in the bottom of the pan at all times. Step 3 Cook until the top of the pancake starts to bubble and edges turn browned and crisp, 2 to 3 minutes. Use a spatula to flip each pancake. The cooked surface should be very crispy, with a dark ring around the edge. Cook until the second side is browned and crisp, 2 to 3 minutes. Repeat to cook the remaining pancakes, adding more clarified butter as needed. Step 4 Serve immediately with pats of salted butter, if desired, and maple syrup. If making a large batch, cooked pancakes can be kept warm on a wire rack set in a rimmed metal baking sheet in a 300-degree oven. The batter can be made ahead and refrigerated overnight. You might need to increase cook time on the pancakes by 1 minute or so on each side. Leave a Private Note on this recipe and see it here.
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Step 1 Make the cake: Heat oven to 325 degrees and place a rack in the center. Butter and line the bottoms of two 8-inch round cake pans with parchment paper. Butter the parchment. Flour the pans, coating the bottom and sides, then tap out any excess flour. Step 2 In a medium bowl, whisk together flour, baking powder, baking soda and salt. In a small bowl, stir together milk and vanilla. Step 3 In a mixer fitted with the paddle attachment, beat butter and sugar at medium-high speed until light and fluffy, 2 to 3 minutes. Reduce speed to low and, with mixer running, slowly add egg and whites and beat until smooth and creamy, 1 to 2 minutes. Scrape down bowl. Add half the flour mixture, then the milk mixture, then the remaining flour mixture, beating to blend after each addition. Scrape down bowl and blend once more. Remove bowl and use a spatula or spoon to mix in sprinkles by hand. Step 4 Divide batter evenly between the pans and bake until the tops are just dry to the touch and a toothpick inserted into the center comes out clean, 20 to 25 minutes. Rotate the pans halfway through baking. Step 5 Transfer pans to a wire rack and let cool completely, at least 4 hours, before turning out. Step 6 Make the frosting: In a mixer fitted with the paddle attachment, beat cream cheese, butter and salt together at medium speed until light and fluffy, 2 minutes. Reduce speed to low and, with mixer running, slowly add sugar and beat until smooth. Beat in vanilla just until incorporated. Taste frosting; you should be able to taste the cream cheese and a little bit of saltiness, as well as the sweetness. If desired, add more salt or sugar. Keep refrigerated. Step 7 Turn cooled cakes out of pans. Using a bread knife, cut off any domed or uneven parts of each cake to make flat surfaces. Place one of the cakes on a platter or a cake stand, cut side up. Using an offset spatula, frost the top. Stack the other layer on top, cut side down. Place remaining frosting on top of the cake and, working from the center outward, frost the top and sides of the cake. Step 8 Place a baking sheet under the platter or cake stand, to catch any fallen sprinkles. Sprinkle the top with a layer of sprinkles. Toss remaining sprinkles at the sides of the cake, or press handfuls of sprinkles gently up the sides to make a thicker coating. Gather fallen sprinkles and repeat until coated to your liking. Step 9 Refrigerate cake 30 minutes or longer to set. Serve cool. Julia's house blend: 2 parts blue, 1 part yellow, 1 part orange, 1 part bright pink. It's best not to use white, light pastel or "natural" sprinkles. They just disappear (poof!) into the cake after baking. Leave a Private Note on this recipe and see it here.
#>                                                                                                  link
#> 1                             https://cooking.nytimes.com/recipes/1018180-perfect-buttermilk-pancakes
#> 2                                                 https://cooking.nytimes.com/recipes/1017409-waffles
#> 3                 https://cooking.nytimes.com/recipes/1017392-pasta-primavera-with-asparagus-and-peas
#> 4                                          https://cooking.nytimes.com/recipes/1019986-olive-oil-cake
#> 5                                  https://cooking.nytimes.com/recipes/1019102-overnight-french-toast
#> 6                                       https://cooking.nytimes.com/recipes/1020697-sugar-cookie-bars
#> 7                                          https://cooking.nytimes.com/recipes/1018490-cinnamon-rolls
#> 8  https://cooking.nytimes.com/recipes/1021062-earl-grey-tea-cake-with-dark-chocolate-and-orange-zest
#> 9                                 https://cooking.nytimes.com/recipes/1020121-chez-ma-tantes-pancakes
#> 10                                  https://cooking.nytimes.com/recipes/1018364-rainbow-sprinkle-cake